Crypto_Agent/pyproject.toml
[project]
name = "crypto-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
